Move services to shared/services folder
