+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnDestroy } from '@angular/core';
-import { LoginService } from './login.service';
+import { LoginService } from './shared/services/login.service';
 
 @Component({
   selector: 'app-root',
