+++ src/app/app.module.ts
@@ -7,15 +7,15 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { UploadService } from './upload-service';
+import { UploadService } from './shared/services/upload-service';
 import { HttpClientModule } from '@angular/common/http';
-import { DragdropDirective } from './dragdrop.directive';
+import { DragdropDirective } from './shared/services/dragdrop.directive';
 import { DashboardComponent } from './dashboard/dashboard.component';
 // Import the module from the SDK
 import { AuthModule } from '@auth0/auth0-angular';
 import { FormsModule, ReactiveFormsModule  } from '@angular/forms';
 import { environment as env } from '../environments/environment';
-import { LoginService } from './login.service';
+import { LoginService } from './shared/services/login.service';
 import { LoginComponent } from './login/login.component';
 import { AdminComponent } from './admin/admin.component';
 import { ReviewComponent } from './review/review.component';
