+++ src/test/login.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { LoginService } from '../app/login.service';
+import { LoginService } from '../app/shared/services/login.service';
 
 describe('LoginService', () => {
   let service: LoginService;
